+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 
 ### Added
 
+- Add script to pull links from all non-archived NOFOs
+
 ### Changed
 
 ### Fixed